Break new class out into a separate file.
IAPWS95 is rather larger than I think is comfortable, and so I took
this opportunity to break a clearly delinated class out into a new
file.
